When hostname is missing, set BAD_CERT_DOMAIN
This fixes a bug with the BadCertHook that we recently introduced: when
no hostname is set, NSS sets SSL_ERRORS_BAD_CERT_DOMAIN, but our custom
callback didn't. This meant it would pass the last seen error back up to
the caller, even though we meant to set this instead.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>