Move JSS to standard Maven layout
We split JSS into a "jss" component and a "native" component in order
to facilitate separate distribution formats for each (.jar vs .so). This
will let us create a parent POM, which builds both .jar and .so, and two
child POMs for each type of distribution. These distributions have
separate subtrees and do not share any common code.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>