alpn: format and supply ALPN data if user provides it
Update the private SSLSocket constructor to accept a String[] of
ALPN protocol names in preferred order.  When given, it will prepare
the ALPN extension data in the required format and pass it along to
the socketConnect method.

The change to publicly expose this capability will occur in a
subsequent commit.