JSSSocketChannel.read: deliver all data from buffer fd
JSSSocketChannel.read() returns 0 when the underlying socket does
not yield any new data.  But in the case of small output buffers,
earlier invocations of read() may result in more decrypted data
sitting in the unwrapped side of the j_buffer, never to be
delievered to the application.

To resolve this issue, continue on to unwrap() even when the
underlying socket does not yield new data.

However, only do this after the handshake has been completed.
Empirically, calling unwrap() at early stages of the handshake with
empty buffers breaks /something/, causing no data to ever be read
from the socket.  (I do not understand the exact cause.)

As part of this change, also refactor how the remote read is
handled.  The choice of which channel to read and handling of
blocking channels is now abstracted behind the remoteRead() method,
which subsumes boundRead().

This commit also addresses a subtle bug.  If the readBuffer contains
more data than the capacity of the j_buffer to which it is written,
and if the dst buffer(s) get filled, then JSSSocketChannel.read()
can terminate with leftover data in readBuffer.  Upon the next
invocation of read(), readBuffer.clear() is called, discarding those
data.  This commit modifies read() to ensure that leftover data in
readBuffer are compacted and preserved.