PR.Read: return data when end of stream reached
Commit 424d55639c51bca18110dcff51fb5df2e2e35a41 introduced a
workaround in PR.Read for a bug in NSS' PR_Read that occurs on some
platforms.  The bug can cause PR_Read to return a smaller amount
than requested, even when more data is available.  The workaround is
to loop until we have either read the requested amount, or we
encounter PR_WOULD_BLOCK_ERROR.

Unfortunatley there is a bug in the workaround: the end of stream
case was not considered.  In this case, the first PR_Read reads less
than the requested amount, and we loop.  The next PR_Read returns 0,
but the "not an error" condition

    PR_GetError() == PR_WOULD_BLOCK_ERROR && read_amount > 0

does not admit the end of stream case (PR_GetError() == 0).  As a
consequence, PR.Read returns an empty byte[] and the data from the
final read before end of stream are lost.

To resolve this issue, break the loop when PR_Read returns <= 0 but
PR_GetError() returns 0.