JSSSocketChannel.read: record when input is closed
When reading the socket channel, no special action is taken when the
SSLEngine.unwrap() returns an SSLEngineResult with status CLOSED.
As a consequence, JSSSocketChannel.read() never returns -1, even
after the sender has closed their send end of the socket and the TLS
close_notify alert has been processed.  Programs that rely on the
read() returning -1 to indicate end of data will loop forever.

To resolve, call shutdownInput() when unwrap() indicates that the
read side of the socket is closed.  This idempotent action causes
subsequent invocations of read() to return -1, as required by the
SocketChannel contract.