Replace RSAPrivateKey with RSAKey
Per discussion with Andrew Hughes in #java, JCA takes RSAPrivateKey to
mean that the private exponent is extractable. Because the JCA is
careful to use Key most places, we might get away with making
PK11RSAPrivateKey not implement RSAPrivateKey and instead only
implement RSAKey -- signalling to other providers that the private
exponent isn't extractable.

This patch requires careful testing with SunJSSE and Dogtag PKI to
ensure we don't rely on PK11RSAPrivateKey implementing RSAPrivateKey
somewhere.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>