Convert KeyPairGeneratorSpi.Usage into enumeration
The KeyPairGeneratorSpi.Usage class has been converted into
an enumeration that defines the same elements. These elements
store the values of the corresponding PKCS #11 constants.

The getVal() method has been deprecated, but it will continue
to work as before for backward compatibility.

The opFlags and opFlagsMask fields in PK11KeyPairGenerator
have been converted to long to match the PKCS #11 constants.
The values will be down-casted to int whenever needed.

The opFlagForUsage array in PK11KeyPairGenerator has become
redundant since the PKCS #11 constants can be obtained from
the enumeration elements, so it has been removed.