Add CA installation test
A Dockerfile has been added to define the builder image and test
runner image for JSS. The builder image is used to install the
build dependencies and then build JSS. The test runner image is
used to install JSS with its runtime dependencies then run tests.

A new test has been added to install DS and CA with the newly
built JSS packages.

Some scripts have been added to initialize the test runner, to
create and remove DS, and to save the test artifacts.