Move pkcs11check into Fedora 33 test
Currently the PKCS #11 constants are only tested on Rawhide
and it's failing, but it's actually working fine on Fedora 33.
For now the test has been moved into the Fedora 33 test so the
CI will pass, but later the tests will need to be expanded to
include newer Fedora versions. That way the PKCS #11 constants
will be tested separately on each platform, so it will be
easier to isolate the problem.