Move CapabilitiesList into tests/ directory
By moving CapabilitiesList, we can take advantage of the existing build
system for building and running CapabilitiesList.

The downside of this approach is that we still need all libraries JSS
depends on to be present for running.

This is made easy by run_test.sh (templated under build/):

    cd build && ./run_test org.mozilla.jss.tests.CapabilitiesList

Note that this gives the warning:

    WARNING: JSS provider is already installed

In order to suppress this warning, the line:

    export JAVA_SECURITY_CFG="...some/path..."

can be overwritten with the empty string:

    export JAVA_SECURITY_CFG=""

and the utility run. However, JSS will not automatically load as a
result of the java.security provider properties listing.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>