Update version number to 5.0.0-alpha1
Recently JSS 4.9 was modified to require Java 11. However,
older platforms that use JSS 4.8 might have other libraries
that will not work with Java 11, so it will not be able to
be upgraded to JSS 4.9.

To address the problem, the current JSS 4.9 has been renamed
into JSS 5.0 which will only work on newer platforms, and a
new JSS 4.9 branch will be created from JSS 4.8 to provide
updates for older platforms.

The binaries have also been renamed to jss.jar and libjss.so
to simplify future upgrades.