Add support for EventListeners in JSSParameters
Here we take the approach of allowing any EventListener and storing it,
accepting it potentially for future use. Right now, we're adding support
for both SSLHandshakeCompletedListener and SSLSocketListener. The former
is the parent class of the latter, so since we need the latter, we might
as well implement both. If future event listeners are necessary, it is
suggested that they should derive directly from EventListener and not
from SSLHandshakeCompletedListener.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>