Add event listeners to JSSEngine interface
Implements top-level interfaces to accept EventListeners in JSSEngine
and adds internal firing helpers for derivative JSSEngine
implementations to call. This abstracts over the differences in classes
and allows the specific event to be passed into the listener, ignoring
unknown listeners.

Note that no filtering of listeners has been done at this time. This
means that a unknown event listener type could be passed into JSSEngine
and it'll be accepted but ignored (and thus, never fire).

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>