Add JSS symkey library
To consolidate Java API for NSS the PKI symkey library has been
imported into JSS with the following changes:

- the package name has been renamed to org.mozilla.jss.symkey
- the JAR file has been renamed to jss-symkey.jar
- the shared library file has been renamed to libjss-symkey.so

The code loading the library in SessionKey has been simplified
as well. In the future the JAR file and the shared library file
might be merged into jss.jar and libjss.so.

https://github.com/dogtagpki/pki/issues/1368