Fix Bug 2013674 - JSS cannot be properly initialized after using another NSS-backed security provider
The solution the the bug is to provide a way to initialize JSS such that the underlying nss system is initialized as a context
instead of a full NSS initialization. This allows jss / nss to create it's own nss init scenario. This could be of use where
the main process has already inited nss with a set of params and properties. Allowing say a library loaded into the man process the ability
to do a context based nss init, allows the library to set it's own nss params and not inherit those from the main process.
For instance if a main process has initialized nss to have a read only database, the context init will allow the library to load an nss
context with a read write database.

This is accomplished in a very simple manner with respect to jss. Right now jss allows one singleton instance to be created for a program.
This fix allows the caller to choose to have the singleton to be inited either as a context or a full init. Theoretically the nss context
concept would allow for multiple contexts to be created,but for simplicity, JSS will only allow one such context or full nss init.

The original behavior of doing a full NSS init beneath JSS persists as the default. In order to do a context init the following sample code will suffice:

CryptoManager.initializeWithContext(databaseDir);

manager = CryptoManager.getInstance();
...
...
manager.shutdown().

Notice that  an explicit call to shutdown is recommended because the NSS context inited must be destroyed at the native code level, in oder to not interfere with the main program's nss instance.

If one wants to explicitly specify the Initialization  values, the following call can be used as well:

InitializationValues vals =
                    new InitializationValues(args[0],
                            "", "", "secmod.db");

CryptoManager.initializeWithContext(vals);
CryptoManager manager = CryptoManager.getIntance();

...
...

manager.shutdwown();

Note: That as of this patch , there is no way to call for a context init when using the JCA provider interface. That can be a future improvement.