Fix the compilation error on Sonarclod with maven (#872)
After the PR #869 maven build does not work if the CMake build is not
execute before. This to properly generate the header for the native
code.

Since Sonarcloud is configured with maven build and it does not check c
code for the moment, the native module has been disabled for the code
analysis.