Fix Azure pipelines
Due to recent changes in Azure or Fedora it's no longer possible
to run tests in Azure with the standard Fedora image.

To fix the problem, the Dockerfile has been updated to define a
new Fedora/CentOS image that provides a systemd service. The test
job in Azure has been split into two jobs: one for Fedora/CentOS
and one for Debian/Ubuntu. The Fedora/CentOS job will build the
image, create a container with the image, then run the tests with
it, whereas the Debian/Ubuntu job is basically unchanged.