Deprecate RSAParameterSpec.getKeySize
The class RSAParameterSpec has been modified to extend the standard `java.security.spec.RSAKeyGenParameterSpec` but the method `getKeySize()` has a different capitalisation in the superclass so this was kept for compatibility.

However, all the *JSS* code has been updated to use the superclass method
and there is no use in other *PKI* tools so the old method is not used
and can be removed in the future.

The commit modifying the RSAParameterSpec extending the standard class
is:

https://github.com/dogtagpki/jss/commit/b995847206e21b696c697f1551b0ee61e17f9497