Publish to GitHub Maven registry
A new job has been added to build JSS with Maven and publish
the artifacts to GitHub Maven registry. Currently the native
and symkey modules have to be disabled due to a build issue.
The group ID and artifact ID have been renamed to follow the
more commonly used pattern.