Install both jss.jar and jss-base.jar
The final location of jss-base.jar can change depending on the
build system (local rpmbuild vs COPR), so it's not possible
to create reliable links to the file.

To fix the problem the RPM spec file has been updated to install
jss.jar (in addition to jss-base.jar) at a more consistent
location so that the links will work properly.