Add IPs address to the session
The SSLEngine session "JSSSession" has been extended to container the IP
addresses of the client and the server. These are used for the audit and
have not other use in the protocol. By design the SSLEngine should be
unaware of the underlying communication but this is need to keep the
original audit format required for the certification.