Update exception handling in PFX.verifyAuthSafes()
Previously the PFX.verifyAuthSafes() discarded the original
exception and simply returned a generic error message which
was usually insufficient to determine the cause of the issue.

To help troubleshooting, the PFX.verifyAuthSafes() has been
modified to throw the exception so the original error message
and the stack trace are preserved.