Enable certification verification using CRL-DP
Currently certificates can be validated only using OCSP with a configured
responder or using the AIA certificate extension. If the responder
cannot be used verification is not possible. This is the case for the
startup certificates of the responder.

The new policy add a verification using the CRL-DP extension defined in
the certificate. If AIA extension is defined it has precedence over the
CRLs and if no other check are performed.
This new method takes place when OCSP is configured without a default
responder and the PKIX verification method is adopted (with the policy
OCSP_LEAF_AND_CHAIN_POLICY).

At least a verification method, defined in the certificate, has to return
success to accept the certificate.