Add revocation check fallback for PKIX based verification
In case of CRLDP and AIA extensions available in the certificate the
PKIX verification flags do not allow to implement a fallback mechanism
because the fallback remain active also for certificate with only one
extension with unexpected behaviour.

A more dynamic approach is introduced verifying the presence of the
CRLDP extension and setting the flags accordingly.