Fix test 17 error on build
With the nss update to >= 3.101 the nss context gets initialised after
the CryptoManager instance is retrieved so this step is added before
the test.

Additional, the ssl option test for get and set has been modified to
enable/disable ssl cache since the ssl3 option is policy dependent and do not
always work.