Add cert approval callback for JSSTrustManager
Previously the JSSTrustManager would throw an exception to
reject a cert if it found an issue in the cert. The code has
been changed to collect all issues, then pass them into an
optional callback object. If the callback approves it, the
cert will be considered trusted. Otherwise, or if there is
no callback, the code will throw an exception based on the
first issue.