JSSTrustManager check revokation  only if trusted issuer
If the certificate is a trusted peer there is not check of the issuer
and/or the certificate.

If the issuer is not available, or it is a trusted peer, the revocation
is not verified because it cannot verify the full chain in any case.
The issue is already reported.