Fix write problem with blocking channel
JSSSocket can be used with blocking and non blocking channel.
When used with blocking channel the write has to send all data in a
single request but if the sending buffer is full only that is
sent and the remaining data will be send in following write operations.

This commit fix the write adding consecutive operations if the buffer is full
in order to send all application data.