Make JSSSocket close() idempotent
Invoking multiple times the close() method generate an exception because
it calls the init() which tries to access closed channels.

The close() method is defined in Closeable interface and it has to be
idempotent so it should be possible to call again without errors.

This is a problem with third party libraries not checking if the socket
has been closed but call multiple times the close() method generating
error.