Add EC public key generation from spec
Public key generation has been updated to allow the creation of ECPublicKey
from the ECPublicKeySpec.
The code was requiring the encoded public key to generate a new key to use
for signing verification but this is a limitation when the encoded key is
not available, as for JWS validation using ES256 algorithm.