Jss missing pqg algs v5.8 (#1054)
* fix build failure in dist-git centos caused by udpated NSS version with new pqc defs int auth_alg_defs[] (#1052)

jss.spec changes will be needed for dist-git.

For reference, this change corresponds to an NSS patch that currently exists on CentOS 10 only:
https://gitlab.com/redhat/centos-stream/rpms/nss/-/blob/c10s/nss-3.112-add-ml-dsa-ssl-support.patch?ref_type=heads#L965

* Adding option for ENABLE_NSS_VERSION_PQC_DEF (#1053)

Note that we still need jss.spec to differentiate NSS versions to decide whether to set ENABLE_NSS_VERSION_PQC_DEF to on or off.

Fix as part of the build task:
https://issues.redhat.com/browse/IDM-2360