Add all ML-DSA variants algorithms
PKI code requires to explicitly specify the signature algorithms so the
generic "ML-DSA" algorithm is replaces with its variants and this is
just a link to the default "ML-DSA-65".

Additionally, KeyType and Cert have been modified to handle ML-DSA
variants so certificates can be generated in JSS using NSS code only for
the signature.